// AngularClient/src/app/api/generated/model/accountingDatasetInfo.ts
/**
 * FinancesApi
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: v1
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { DatasetState } from './datasetState';


export interface AccountingDatasetInfo { 
    state?: DatasetState;
    lastCloseDate?: string;
    message?: string | null;
}